Refactor: Relocate Favorites to Settings and revamp navigation
This commit implements the following changes:

1.  **Favorites Tab Relocated:**
    *   The "Favorites" tab functionality has been moved into the Settings section.
    *   A new page at `/settings/favorites` now displays your favorite produce.
    *   Links to "My Favorites" have been added to the Settings sheet in the mobile bottom navigation and to the Settings group in the desktop sidebar.

2.  **Navigation Changes:**
    *   The former "Favorites" tab in the main navigation (mobile bottom nav and desktop sidebar) has been renamed to "Search".
    *   Its icon has been updated from a heart to a search icon.
    *   This "Search" tab now links to `/search`.

3.  **Home Screen Content Moved:**
    *   The entire content and layout of the original Home screen (previously at `/`) has been moved to the new "Search" tab page (`/search`). This page retains all search, filter, and produce display functionalities.

4.  **Home Screen Blanked:**
    *   The main Home screen (at `/`) is now a blank page, as per requirements.

These changes restructure the app's navigation to better categorize features, placing favorites within user settings and dedicating the prominent former "Favorites" slot to the primary search/discovery functionality.
